Refactor instance model imports for consistency.
Updated import paths for `InstanceStatus` and `IInstance` to point to the correct `models/instance.model` location. This ensures alignment with the updated project structure and improves maintainability.
